delete: remoção de metodos não utilizados e comentários desnecessários
