+++ sysagua-app/src/main/java/edu/pies/sysaguaapp/services/produto/ProductCategoryService.java
@@ -1,6 +1,6 @@
-package edu.pies.sysaguaapp.services;
+package edu.pies.sysaguaapp.services.produto;
 
-import edu.pies.sysaguaapp.models.ProductLine;
+import edu.pies.sysaguaapp.models.ProductCategory;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -10,17 +10,17 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.util.List;
 
-public class ProductLineService {
-    private static final String BASE_URL = "http://localhost:8080/products/line";
+public class ProductCategoryService {
+    private static final String BASE_URL = "http://localhost:8080/products/category";
     private final HttpClient httpClient;
     private final ObjectMapper objectMapper;
 
-    public ProductLineService() {
+    public ProductCategoryService() {
         this.httpClient = HttpClient.newHttpClient();
         this.objectMapper = new ObjectMapper();
     }
 
-    public List<ProductLine> buscarLinhas() throws Exception {
+    public List<ProductCategory> buscarCategorias() throws Exception {
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(URI.create(BASE_URL))
                 .GET()
@@ -30,9 +30,9 @@ public class ProductLineService {
         HttpResponse<String> response = httpClient.send(request, HttpResponse.BodyHandlers.ofString());
 
         if (response.statusCode() == 200) {
-            return objectMapper.readValue(response.body(), new TypeReference<List<ProductLine>>() {});
+            return objectMapper.readValue(response.body(), new TypeReference<List<ProductCategory>>() {});
         } else {
-            throw new Exception("Erro ao buscar linhas: " + response.body());
+            throw new Exception("Erro ao buscar categorias: " + response.body());
         }
     }
 }
