+++ sysagua-app/src/main/java/edu/pies/sysaguaapp/services/produto/ProdutoService.java
@@ -1,17 +1,13 @@
-package edu.pies.sysaguaapp.services;
+package edu.pies.sysaguaapp.services.produto;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import edu.pies.sysaguaapp.models.Fornecedor;
 import edu.pies.sysaguaapp.models.Produto;
 
 import java.net.URI;
-import java.net.URLEncoder;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
 import java.util.List;
 
 public class ProdutoService {
